// src/types/MessageRole.ts
// Code generated by schema2code at 2026-03-07 19:34:28. DO NOT EDIT.


/**
 * Role of a participant in a chat conversation
 */
export type MessageRole = 'user' | 'assistant' | 'system' | 'tool' | 'agent' | 'observer';

/**
 * Constant values for MessageRole
 */
export const MessageRoleValues = {
  /** user */
  USER: 'user',
  /** assistant */
  ASSISTANT: 'assistant',
  /** system */
  SYSTEM: 'system',
  /** tool */
  TOOL: 'tool',
  /** agent */
  AGENT: 'agent',
  /** observer */
  OBSERVER: 'observer'
} as const;